порты и адаптеры

+++ src/Ebay/Server.Application/Abstractions/Models/ShippingRates/ShippingRateWithoutCurrency.cs
@@ -1,3 +1,3 @@
 ﻿namespace Server.Application.Abstractions.Models.ShippingRates;
 
-public record ShippingRateWithoutCurrency(int WeightFrom, int WeightTo, double Price);
+public record ShippingRateWithoutCurrency(int WeightFrom, int WeightTo, int Price);
